Refactor virtual life integration
