+++ app/core/virtual_life.py
@@ -334,68 +334,3 @@ class VirtualLifeManager:
         except Exception as e:
             self.logger.error(f"Ошибка получения планов: {e}")
             return []
-
-# ДОБАВИТЬ В companion.py:
-
-class RealisticAICompanion:
-    def __init__(self, config: Dict[str, Any]):
-        # ... существующий код ...
-        
-        # НОВОЕ: Система виртуальной жизни
-        self.virtual_life = VirtualLifeManager(
-            db_path=config.get('database', {}).get('path', 'data/companion.db'),
-            character_loader=character_loader
-        )
-        
-        # Добавляем проверку активностей в планировщик
-        self.scheduler.add_job(
-            self.update_virtual_life,
-            IntervalTrigger(minutes=1),  # Проверяем каждую минуту
-            id='virtual_life_update'
-        )
-    
-    async def update_virtual_life(self):
-        """Обновляет виртуальную жизнь персонажа"""
-        try:
-            changes = self.virtual_life.check_and_update_activities()
-            
-            if changes["status_changed"]:
-                # Персонаж начал новую активность
-                if changes["activity_started"]:
-                    activity = changes["activity_started"]
-                    await self._notify_activity_start(activity)
-                
-                # Персонаж закончил активность
-                if changes["activity_ended"]:
-                    activity = changes["activity_ended"]
-                    await self._notify_activity_end(activity)
-                    
-        except Exception as e:
-            self.logger.error(f"Ошибка обновления виртуальной жизни: {e}")
-    
-    async def _notify_activity_start(self, activity: VirtualActivity):
-        """Уведомляет о начале активности"""
-        # Генерируем сообщение о начале активности
-        messages = [
-            f"Кстати, я сейчас {activity.description}! ✨",
-            f"Буду {activity.activity_type} до {activity.end_time.strftime('%H:%M')}",
-            "Но ты всегда можешь мне писать! 💕"
-        ]
-        
-        # Отправляем если есть активные пользователи
-        if hasattr(self, 'allowed_users') and self.allowed_users:
-            current_state = await self.optimized_ai.get_simple_mood_calculation(
-                self.psychological_core
-            )
-            
-            for user_id in self.allowed_users:
-                try:
-                    await self.send_telegram_messages_with_timing(
-                        chat_id=user_id,
-                        messages=messages,
-                        current_state=current_state
-                    )
-                except Exception as e:
-                    self.logger.error(f"Ошибка отправки уведомления активности: {e}")
-
-# Добавить в AI промпт контекст виртуальной жизни
